use int instead of floats for scores
nn output is scaled by 64*127 and no tanh is applied.

Score of bread_engine_1.0.2 vs bread_engine_1.0.0: 413 - 472 - 302  [0.475] 1187
...      bread_engine_1.0.2 playing White: 268 - 189 - 136  [0.567] 593
...      bread_engine_1.0.2 playing Black: 145 - 283 - 166  [0.384] 594
...      White vs Black: 551 - 334 - 302  [0.591] 1187
Elo difference: -17.3 +/- 17.1, LOS: 2.4 %, DrawRatio: 25.4 %
SPRT: llr -2.94 (-99.9%), lbound -2.94, ubound 2.94

+++ dependencies/extern/chess.hpp
@@ -1178,10 +1178,10 @@ class Move {
 
     /// @brief Set the score for a move. Useful if you later want to sort the moves.
     /// @param score
-    constexpr void setScore(float score) noexcept { score_ = score; }
+    constexpr void setScore(int score) noexcept { score_ = score; }
 
     [[nodiscard]] constexpr std::uint16_t move() const noexcept { return move_; }
-    [[nodiscard]] constexpr float score() const noexcept { return score_; }
+    [[nodiscard]] constexpr int score() const noexcept { return score_; }
 
     constexpr bool operator==(const Move &rhs) const noexcept { return move_ == rhs.move_; }
     constexpr bool operator!=(const Move &rhs) const noexcept { return move_ != rhs.move_; }
@@ -1195,7 +1195,7 @@ class Move {
 
    private:
     std::uint16_t move_;
-    float score_;
+    int score_;
 };
 
 inline std::ostream &operator<<(std::ostream &os, const Move &move) {
