+++ dependencies/transposition_table.hpp
@@ -16,7 +16,7 @@ enum class TFlag: uint8_t {
 // side to move is not stored in the transposition table as it is in the zobrist hash
 struct TEntry {
     uint64_t zobrist_hash = 0; // 8 bytes
-    float evaluation = 0; // 4 bytes
+    int evaluation = 0; // 4 bytes
     uint16_t best_move = 0; // 2 bytes
     uint8_t depth_tflag = 0; // 1 byte  -> contains depth: 6 bits (max 63), flag: 2 bits (4 values)
     uint8_t move_number = 0; // 1 byte
@@ -33,7 +33,7 @@ struct TEntry {
 
     TEntry(){};
 
-    TEntry(uint64_t zobrist, float eval, int depth, chess::Move move, TFlag flag, uint8_t move_number):
+    TEntry(uint64_t zobrist, int eval, int depth, chess::Move move, TFlag flag, uint8_t move_number):
             zobrist_hash(zobrist),
             evaluation(eval),
             best_move(move.move()),
@@ -48,7 +48,7 @@ class TranspositionTable {
 
     void allocateMB(int new_size);
 
-    void store(uint64_t zobrist, float eval, int depth, chess::Move move, TFlag flag, uint8_t move_number);
+    void store(uint64_t zobrist, int eval, int depth, chess::Move move, TFlag flag, uint8_t move_number);
     void store(TEntry entry);
 
     TEntry* probe(bool& is_hit, uint64_t zobrist);
